adding settings
