+++ src/example.cpp
@@ -5,8 +5,8 @@
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
-#include "../utils/Utils.hpp"
-#include "../utils/Sphere.hpp"
+#include <Utils.hpp>
+#include <Sphere.hpp>
 
 using namespace std;
