+++ src/program_5_1_1.cpp
@@ -7,7 +7,7 @@
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
-#include "../utils/Utils.hpp"
+#include <Utils.hpp>
 
 using namespace std;
