+++ src/stlHeaders.hpp
@@ -1,0 +1,9 @@
+#pragma once 
+
+#include <vector>
+#include <string>
+#include <iostream>
+#include <fstream>
+#include <sstream>
+#include <cmath>
+
