+++ utils/CMakeLists.txt
@@ -6,17 +6,23 @@ find_package(glm REQUIRED)
 find_package(OpenGL REQUIRED)
 find_package(SOIL2 REQUIRED)
 
-#target_link_libraries(${SOURCE_FILE_NAME} glad glfw glm::glm soil2 ${OPENGL_LIBRARIES} utilities)
-
 set(MAIN_SOURCE_FILES 
-    ${CMAKE_CURRENT_SOURCE_DIR}/Sphere.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Torus.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Utils.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/Sphere.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/Torus.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/Utils.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/modelImporter.cpp
 )
+
 
 add_library(utilities SHARED ${MAIN_SOURCE_FILES})
 
-target_include_directories(utilities PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+# Directorios de inclusión públicos
+target_include_directories(utilities PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include/public)
+
+# Directorios de inclusión privados
+target_include_directories(utilities PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include/private)
+
+target_precompile_headers(utilities PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include/private/stlHeaders.hpp )
 
 target_link_libraries(utilities PRIVATE
     glad
