+++ utils/include/private/stlHeaders.hpp
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef STLHEADERS_HPP
+#define STLHEADERS_HPP
 
 #include <string>
 #include <iostream>
@@ -7,3 +8,5 @@
 #include <vector>
 #include <fstream>
 #include <sstream>
+
+#endif
