+++ utils/include/public/Utils.hpp
@@ -1,7 +1,9 @@
+#pragma once
+
 #include <glad/glad.h>
 #include <GLFW/glfw3.h>
 #include <GLFW/glfw3native.h>
-#include "stlHeaders.hpp"
+#include "../private/stlHeaders.hpp"
 #include <soil2/SOIL2.h>
 #include <glm/glm.hpp>
 #include <glm/gtc/type_ptr.hpp>
