+++ utils/src/Utils.cpp
@@ -1,5 +1,5 @@
 //#include <glad/glad.h>
-#include "Utils.hpp"
+#include <Utils.hpp>
 
 using namespace std;
