+++ utils/src/modelImporter.cpp
@@ -1,4 +1,4 @@
-#include"modelImporter.hpp"
+#include <modelImporter.hpp>
 ModelImporter::ModelImporter() {}
 
 void ModelImporter::parseObjFile(const char* path) {
